requirements.txt
tensorflow>=2.11,<3.0
numpy
pandas
matplotlib
Pillow
streamlit
scikit-learn
